requirements.txt
flask
tensorflow
pillow
numpy
requests
instaloader
werkzeug
gunicorn